Fix some intermittent test failures
I was seeing these fail when Faker generated job titles for jobs that should match that were the same or similar enough to ones that shouldn't match. So this sets them explicitly so that we don't have intermittent failures. There might be more instances of this in the tests, though.